+++ x/bridge/types/constants.go
@@ -1,0 +1,8 @@
+package types
+
+const (
+	DefaultBitcoinChainName = "bitcoin"
+	DefaultBitcoinDenomName = "btc"
+	DefaultBitcoinExponent  = 10 // TODO: decide
+	DefaultVotesNeeded      = 1
+)